docker_build/src/model/analyzer.py
"""
LLaVA分析器模块，用于分析图表和生成交易信号
"""
from typing import Dict, Any, Optional, List
import logging
from pathlib import Path
import torch
from PIL import Image
import json
import os

from config.config import LLAVA_MODEL_PATH, LLAVA_PROMPT_TEMPLATE
from src.utils.logger import setup_logger

logger = setup_logger('llava_analyzer')

class LLaVAAnalyzer:
    """LLaVA分析器类，用于分析图表和生成交易信号"""
    
    def __init__(self):
        """初始化LLaVA分析器"""
        try:
            logger.info("Initializing LLaVA analyzer...")
            self.model_path = LLAVA_MODEL_PATH
            self.prompt_template = LLAVA_PROMPT_TEMPLATE
            self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
            
            # 加载模型
            self._load_model()
            logger.info("LLaVA analyzer initialized successfully")
            
        except Exception as e:
            logger.error(f"Failed to initialize LLaVA analyzer: {e}")
            raise
            
    def _load_model(self):
        """加载LLaVA模型"""
        try:
            logger.info(f"Loading LLaVA model from: {self.model_path}")
            
            # TODO: 实现模型加载逻辑
            # 这里需要根据实际使用的LLaVA模型实现加载逻辑
            
            logger.info("LLaVA model loaded successfully")
            
        except Exception as e:
            logger.error(f"Failed to load LLaVA model: {e}")
            raise
            
    def analyze_chart(self, image_path: str) -> Dict[str, Any]:
        """
        分析图表并生成交易信号
        
        Args:
            image_path: 图表图片路径
            
        Returns:
            分析结果字典
        """
        try:
            logger.info(f"Analyzing chart: {image_path}")
            
            # 加载图片
            image = self._load_image(image_path)
            if image is None:
                raise ValueError(f"Failed to load image: {image_path}")
                
            # 生成提示词
            prompt = self._generate_prompt()
            
            # 运行模型推理
            result = self._run_inference(image, prompt)
            
            # 解析结果
            analysis = self._parse_result(result)
            
            logger.info(f"Chart analysis completed: {analysis}")
            return analysis
            
        except Exception as e:
            logger.error(f"Failed to analyze chart: {e}")
            return {
                'error': str(e),
                'success': False
            }
            
    def _load_image(self, image_path: str) -> Optional[Image.Image]:
        """加载图片"""
        try:
            if not os.path.exists(image_path):
                logger.error(f"Image file not found: {image_path}")
                return None
                
            image = Image.open(image_path)
            return image
            
        except Exception as e:
            logger.error(f"Failed to load image: {e}")
            return None
            
    def _generate_prompt(self) -> str:
        """生成提示词"""
        return self.prompt_template
        
    def _run_inference(self, image: Image.Image, prompt: str) -> str:
        """
        运行模型推理
        
        Args:
            image: 输入图片
            prompt: 提示词
            
        Returns:
            模型输出文本
        """
        try:
            logger.info("Running model inference...")
            
            # TODO: 实现模型推理逻辑
            # 这里需要根据实际使用的LLaVA模型实现推理逻辑
            
            # 临时返回模拟结果
            mock_result = {
                "sequence_type": "上升",
                "key_points": [
                    {"price": 15000, "time": "2024-03-01 10:00:00"},
                    {"price": 15100, "time": "2024-03-01 10:30:00"},
                    {"price": 15200, "time": "2024-03-01 11:00:00"}
                ],
                "market_analysis": "市场呈现上升趋势，成交量放大",
                "trading_suggestion": "建议在回调时买入",
                "confidence": 85
            }
            
            return json.dumps(mock_result, ensure_ascii=False)
            
        except Exception as e:
            logger.error(f"Inference failed: {e}")
            raise
            
    def _parse_result(self, result: str) -> Dict[str, Any]:
        """
        解析模型输出结果
        
        Args:
            result: 模型输出文本
            
        Returns:
            解析后的结果字典
        """
        try:
            # 解析JSON结果
            data = json.loads(result)
            
            # 构建分析结果
            analysis = {
                'success': True,
                'sequence_type': data['sequence_type'],
                'key_points': data['key_points'],
                'market_analysis': data['market_analysis'],
                'trading_suggestion': data['trading_suggestion'],
                'confidence': data['confidence'] / 100.0,  # 转换为0-1范围
                'timestamp': data['key_points'][-1]['time']  # 使用最后一个关键点的时间
            }
            
            return analysis
            
        except Exception as e:
            logger.error(f"Failed to parse result: {e}")
            return {
                'success': False,
                'error': str(e)
            }
            
    def cleanup(self):
        """清理资源"""
        try:
            # TODO: 实现资源清理逻辑
            logger.info("Cleaning up resources...")
            
        except Exception as e:
            logger.error(f"Cleanup failed: {e}") 